Add meta and index manager
quite hacky to use static methods but it is the fastest way ...
